+++ lib/kiwi_codec/decoder.ex
@@ -41,7 +41,7 @@ defmodule KiwiCodec.Decoder do
 
       id ->
         field = Map.get(fields_by_id, id) || raise_unknown_field(struct, id)
-        {value, tail} = decode_field_value(rest, field, struct.__struct__)
+        {value, tail} = decode_wire_field(rest, field, struct.__struct__)
         decode_message(tail, Map.put(struct, field.name, value), metadata)
     end
   end
@@ -53,47 +53,48 @@ defmodule KiwiCodec.Decoder do
   defp decode_struct_fields([], struct, binary, _module), do: {struct, binary}
 
   defp decode_struct_fields([field | fields], struct, binary, module) do
-    {value, rest} = decode_field_value(binary, field, module)
+    {value, rest} = decode_wire_field(binary, field, module)
     decode_struct_fields(fields, Map.put(struct, field.name, value), rest, module)
   end
 
-  defp decode_field_value(binary, %Field{} = field, module) do
-    do_decode_field_value(binary, field)
+  defp decode_wire_field(binary, %Field{} = field, module) do
+    do_decode_wire_field(binary, field)
   rescue
     error in [KiwiCodec.DecodeError, ArgumentError, FunctionClauseError, MatchError] ->
       raise_decode_field_error(module, field, error)
   end
 
-  defp do_decode_field_value(binary, %Field{repeated?: true, type: :byte}) do
+  defp do_decode_wire_field(binary, %Field{repeated?: true, type: :byte}) do
     Wire.decode_byte_array(binary)
   end
 
-  defp do_decode_field_value(binary, %Field{repeated?: true} = field) do
+  defp do_decode_wire_field(binary, %Field{repeated?: true} = field) do
     {length, rest} = Varint.decode_uint(binary)
-    decode_repeated(length, field.type, rest, [])
+    decode_repeated_wire_type(length, field.type, rest, [])
   end
 
-  defp do_decode_field_value(binary, %Field{} = field) do
-    decode_scalar(field.type, binary)
+  defp do_decode_wire_field(binary, %Field{} = field) do
+    decode_wire_type(field.type, binary)
   end
 
-  defp decode_repeated(0, _type, binary, acc), do: {Enum.reverse(acc), binary}
+  defp decode_repeated_wire_type(0, _type, binary, acc), do: {Enum.reverse(acc), binary}
 
-  defp decode_repeated(count, type, binary, acc) do
-    {value, rest} = decode_scalar(type, binary)
-    decode_repeated(count - 1, type, rest, [value | acc])
+  defp decode_repeated_wire_type(count, type, binary, acc) do
+    {value, rest} = decode_wire_type(type, binary)
+    decode_repeated_wire_type(count - 1, type, rest, [value | acc])
   end
 
-  defp decode_scalar({:enum, module}, binary) do
+  defp decode_wire_type({:enum, module}, binary) do
     {value, rest} = Varint.decode_uint(binary)
     {module.key(value), rest}
   end
 
-  defp decode_scalar(type, binary)
+  defp decode_wire_type(type, binary)
        when type in [:bool, :byte, :float, :int, :int64, :string, :uint, :uint64],
        do: Wire.decode(type, binary)
 
-  defp decode_scalar(module, binary) when is_atom(module), do: decode_from_module(binary, module)
+  defp decode_wire_type(module, binary) when is_atom(module),
+    do: decode_from_module(binary, module)
 
   defp transform(message, module) do
     case module.transform_module() do